clean_city-main/clean_city/serializers/checklist_serializer.py
from clean_city.models import CheckList
from rest_framework import serializers


class CheckListSerializer(serializers.ModelSerializer):
    class Meta:
        model = CheckList
        fields = "__all__"
