clean_city-main/requirements.txt
asgiref==3.3.4
attrs==21.2.0
autopep8==1.6.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
cloudinary==1.28.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.21.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==4.2.1
MarkupSafe==2.0.1
packaging==20.9
Pillow==9.0.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
tk==0.1.0
tk-tools==0.14.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-decouple==3.4
pytz==2021.1
PyYAML==6.0
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
zipp==3.7.0
